Move ExprPurifier to own module and remove unnecesarry fields vector
