+++ prusti-viper/src/encoder/snapshot/mod.rs
@@ -1,0 +1,5 @@
+
+
+pub use self::purifier::ExprPurifier;
+
+mod purifier;
